% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parametric_function.R
\name{derivate_Gaussian_2}
\alias{derivate_Gaussian_2}
\title{the derivate of 2-order Gaussian kernel function}
\usage{
derivate_Gaussian_2(center, scatter_point, brand_width)
}
\description{
the derivate of 2-order Gaussian kernel function
}

