% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parametric_function.R
\name{derivate_S1}
\alias{derivate_S1}
\title{derivate_S1 function is the derivate of S_1 function}
\usage{
derivate_S1(center, sample_points, brand_width)
}
\description{
derivate_S1 function is the derivate of S_1 function
}

