% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parametric_function.R
\name{Gaussian_2}
\alias{Gaussian_2}
\title{Functions build for estimating a given non-parametric model}
\usage{
Gaussian_2(center, scatter_point, brand_width)
}
\description{
Functions build for estimating a given non-parametric model
}

