% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parametric_function.R
\name{new_projected_dir}
\alias{new_projected_dir}
\title{new_projected_dir is the function which input residuals, design_matrix,
old projected direction and non_parametric model to output the new
direction generated by the quasi-Newton algorithm.}
\usage{
new_projected_dir(residuals, design_matrix, old_projected_dir, np_model)
}
\description{
@export
}

