% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_function.R
\name{S}
\alias{S}
\title{non-parametric function
#'@include non_parametric_function.R
local linear regression model}
\usage{
S(projection_dir, residuals, design_matrix)
}
\description{
non-parametric function
#'@include non_parametric_function.R
local linear regression model
}

