% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parametric_function.R
\name{S_1}
\alias{S_1}
\title{S_1 function is a part of the derivate of local linear regression function}
\usage{
S_1(center, sample_points, brand_width)
}
\description{
S_1 function is a part of the derivate of local linear regression function
}

