% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parametric_function.R
\name{S_2}
\alias{S_2}
\title{S_2 funciton is another part of the derivate of the local linear regression function}
\usage{
S_2(center, sample_points, brand_width)
}
\description{
@export
}

